feat!: split config.py mega-module into config/ sub-package (#458)
Replace the 1984-line config.py monolith with a config/ package
containing 9 focused sub-modules:

- constants.py: DEFAULT_RISK_FREE_RATE
- optimizer.py: BusinessOptimizerConfig, DecisionEngineConfig
- simulation.py: SimulationConfig, GrowthConfig, DebtConfig, etc.
- manufacturer.py: ManufacturerConfig, IndustryConfig variants
- insurance.py: InsuranceConfig, InsuranceLayerConfig, LossDistributionConfig
- market.py: PricingScenario, MarketCycles, TransitionProbabilities
- reporting.py: OutputConfig, LoggingConfig, ExcelReportConfig
- presets.py: ProfileMetadata, ModuleConfig, PresetConfig, PresetLibrary
- core.py: Config, ConfigV2

All 30 classes are re-exported from config/__init__.py, so every
existing `from ergodic_insurance.config import X` continues to work
with zero changes required in consuming code.

Closes #458
